+++ src/components/Layout/Theme/ToggleColorTheme.jsx
@@ -5,7 +5,7 @@ import { Box, Paper } from "@mui/material";
 import Brightness4Icon from "@mui/icons-material/Brightness4";
 import Brightness7OutlinedIcon from "@mui/icons-material/Brightness7Outlined";
 
-import useThemeStoreSelectors from "../../stores/useThemeStore";
+import useThemeStoreSelectors from "../../../stores/useThemeStore";
 
 const ToggleColorTheme = ({ toggleTheme }) => {
   const theme = useThemeStoreSelectors.use.theme();
