+++ src/components/ReactHookFormComponents/hooks/useLocalStorage.js
@@ -5,7 +5,7 @@ const initialState = (key, initialValue) => {
         const item = window.localStorage.getItem(key)
         return item ? JSON.parse(item) : initialValue
     } catch (error) {
-        console.log(error)
+        console.error(error)
         return initialValue
     }
 }
@@ -17,7 +17,7 @@ const useLocalStorage = (key, initialValue) => {
             setStoredValue(valueToStore)
             window.localStorage.setItem(key, JSON.stringify(valueToStore))
         } catch (error) {
-            console.log(error)
+            console.error(error)
         }
     }
     return [storedValue, setValue]
