--- src/context/Auth/AuthContext.jsx
@@ -1,40 +1,0 @@
-import { createContext, useEffect, useState } from "react";
-
-import PropTypes from "prop-types";
-
-import { auth } from "../../firebase";
-
-import { onAuthStateChanged } from "firebase/auth";
-
-const AuthContext = createContext();
-
-const AuthProvider = ({ children }) => {
-  const [user, setUser] = useState(null);
-  const [isFetching, setIsFetching] = useState(true);
-  useEffect(() => {
-    const unsubscribe = onAuthStateChanged(auth, (user) => {
-      if (user) {
-        setUser(user);
-        setIsFetching(false);
-        return;
-      }
-
-      setUser(null);
-      setIsFetching(false);
-    });
-    return () => unsubscribe();
-  }, []);
-
-  if (isFetching) {
-    return <h2>Loading...</h2>;
-  }
-
-  return (
-    <AuthContext.Provider value={{ user }}>{children}</AuthContext.Provider>
-  );
-};
-AuthProvider.propTypes = {
-  children: PropTypes.node.isRequired,
-};
-
-export { AuthContext, AuthProvider };
